overcooked_ai-master/requirements.txt
# Overcooked AI Requirements
# Install with: pip install -r requirements.txt

# Core dependencies
dill
gymnasium
ipywidgets
numpy
opencv-python
pandas
pygame
scipy
setuptools>=77.0.3
tqdm

# PyTorch for Behavior Cloning
torch>=2.0
tensorboard>=2.14

# JAX for PPO training (CPU version)
jax>=0.4
jaxlib>=0.4
flax>=0.8
optax>=0.2

# Probabilistic Programming (PPL models)
pyro-ppl>=1.8.0
numpyro>=0.13.0

# Experiment tracking (optional)
wandb
sacred

# Visualization
matplotlib
seaborn

# Jupyter support
ipykernel

# Image processing for GIFs
Pillow

